+++ apps/web/src/components/createFormTemplate/FormTemButtons.tsx
@@ -1,0 +1,100 @@
+import { Button, Flex, Text } from '@chakra-ui/react';
+import { useRouter } from 'next/router';
+
+/**
+ * Delete, Back, and Save & Continue buttons at the bottom of form template creation flow.
+ * @param deleteFunction the function to call when delete button is pressed
+ * @param submitLink page router will push on click of 'save & continue'
+ * @param backLink page router will push on click of 'back'
+ * @param disabled whether the 'save & continue' button should be activated
+ */
+export const FormTemButtons = ({
+  deleteFunction,
+  submitLink,
+  backLink,
+  disabled,
+}: {
+  deleteFunction: Function;
+  submitLink: string;
+  backLink: string;
+  disabled: boolean;
+}) => {
+  const router = useRouter();
+  return (
+    <>
+      <Button
+        borderRadius="6px"
+        borderWidth="1.5px"
+        borderStyle={'solid'}
+        borderColor="#E23F40"
+        alignContent={'center'}
+        bgColor={'transparent'}
+        _hover={{
+          bgColor: 'transparent',
+        }}
+        marginLeft="36px"
+      >
+        <Text
+          color="#E23F40"
+          fontWeight="600px"
+          fontSize="18px"
+          lineHeight="22px"
+          onClick={(e) => deleteFunction(e)}
+        >
+          Delete
+        </Text>
+      </Button>
+      <Flex float="right" justifyContent={'space-between'}>
+        <Button
+          borderRadius="6px"
+          borderWidth="1.5px"
+          borderStyle={'solid'}
+          borderColor="#1367EA"
+          alignContent={'center'}
+          bgColor={'transparent'}
+          _hover={{
+            bgColor: 'transparent',
+          }}
+          onClick={() => {
+            router.push(backLink);
+          }}
+        >
+          <Text
+            color="#1367EA"
+            fontWeight="600px"
+            fontSize="18px"
+            lineHeight="22px"
+          >
+            Back
+          </Text>
+        </Button>
+        <Button
+          borderRadius="6px"
+          alignContent={'center'}
+          background={
+            disabled
+              ? 'linear-gradient(0deg, rgba(223, 223, 223, 0.50) 0%, rgba(223, 223, 223, 0.50) 100%), #1367EA;'
+              : '#1367EA'
+          }
+          _hover={{
+            background: 'auto',
+          }}
+          marginLeft="12px"
+          marginRight="36px"
+          onClick={() => {
+            disabled === false ? router.push(submitLink) : null;
+          }}
+        >
+          <Text
+            color="#FCFCFC"
+            fontWeight="600px"
+            fontSize="18px"
+            lineHeight="22px"
+          >
+            Save & Continue
+          </Text>
+        </Button>
+      </Flex>
+    </>
+  );
+};
